Updated FieldName validation logic to now match against expected Regex patterns. Valid/Invalid field names are still generated from a property tree, however the field names are stored as Regex patterns.
Also addressed issue with validating field names that defined an array indexer in the path.

Changed the SortingFields and Filters properties to be nullable to avoid MVC validation superseding Fluent Validation. This prevents the MVC "property is required" error when a request does have any of the properties present in the request.